% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_neighbourhoods.R
\name{assign_neighbourhoods}
\alias{assign_neighbourhoods}
\title{assign_neighbourhoods}
\usage{
assign_neighbourhoods(
  sce,
  k = 25,
  prop = 0.2,
  order = 2,
  filtering = T,
  reducedDim.name,
  k_init = 50,
  d = 30
)
}
\arguments{
\item{sce}{SCE object}

\item{k}{Positive integer, defines how many neighbours to use for the hood assignment}

\item{prop}{Numerical, between 0 and 1, defines fraction of cells from SCE to use for the hoods}

\item{order}{In {1,2}, defines which order of neighbours to use}

\item{filtering}{In {T,F}, defines whether to filter hoods (reduces computing time greatly). Default = TRUE}

\item{reducedDim.name}{defines the slot in reducedDim(sce) to use as embedding for graph construction}

\item{k_init}{Positive integer, defines how many neighbours to use for identifying anchor cells}

\item{d}{Positive integer, defines how many dimensions from reducedDim(sce) to use}
}
\value{

}
\description{
Assign hoods to SCE object
}
