% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test_neighbourhoods.R
\name{de_test_neighbourhoods}
\alias{de_test_neighbourhoods}
\title{de_test_neighbourhoods}
\usage{
de_test_neighbourhoods(
  sce,
  sample_id = "sample",
  condition_id,
  subset_nhoods = NULL,
  covariates = NULL,
  min_n_cells_per_sample = 2,
  gene_selection = c("none", "per_neighbourhood"),
  genes_2_test = rownames(sce),
  min_count = 3,
  output_type = "data.frame",
  plot_summary_stat = FALSE,
  layout = "UMAP",
  BPPARAM = NULL
)
}
\arguments{
\item{sce}{Milo object}

\item{sample_id}{Character specifying which variable should be used as a sample/replica id. Should be in colData(sce)}

\item{condition_id}{Character specifying which variable should be used as a condition id. Should be in colData(sce)}

\item{subset_nhoods}{NULL or character vector specifying the set of neighbourhoods that will be tested for DE}

\item{covariates}{Vector specifying if additional covariates should be passed into experimental design. Default = NULL (no covariates)}

\item{min_n_cells_per_sample}{Positive integer specifying the minimum number of cells per replica to be included in testing. Default = 2}

\item{gene_selection}{In {"none" , "per_neighbourhood"}. If == "per_neighbourhood", it will further discard genes not relevant for this particular neighbourhood.
If == "all", it will perform overall selection and then test each hood across same genes (default). If "none" = no selection based on expression.}

\item{genes_2_test}{Character vector of genes for which we will perform DE testing. Default = rownames(sce). Only applicable if gene_selection == "none".}

\item{min_count}{Positive integer, specifying min.count for gene selection. Default = 3.}

\item{output_type}{In {"data.frame","SCE"} Specifying the output format - either in data.frame or sce (with assays corresponding to logFC, pvals (raw and corrected)); columns correspond to neighbourhoods. Note that default is data.frame, but if number of genes x neighbourhoods combinations is > 10^8 -- the output will be sce.}

\item{plot_summary_stat}{Boolean specifying if we plot Milo neighbourhood plot summarising per neighbourhood whether testing was performed}

\item{layout}{A character indicating the name of the \code{reducedDim} slot in the \code{\linkS4class{Milo}} object to use for layout (default: 'UMAP'). Only relevant if plot_summary_stat == TRUE.}

\item{BPPARAM}{NULL or BPPARAM object}
}
\value{

}
\description{
Tests neighbourhoods for DE + performs per gene correction across neighbourhoods
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
de_stat = de_test_neighbourhoods(sce , condition_id = "type" , gene_selection = "none" , plot_summary_stat = FALSE)
de_stat = convert_de_stat(de_stat)
de_stat = convert_de_stat(de_stat)
}
