% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test_all_hoods.R
\name{de_test_single_hood}
\alias{de_test_single_hood}
\title{de_test_single_hood}
\usage{
de_test_single_hood(
  sce,
  genes = rownames(sce),
  nhoods_sce,
  hood.id,
  sample.id,
  condition.id,
  covariates = NULL,
  min_n_cells_per_sample = 2,
  gene_selection = "all",
  min.count = 5
)
}
\arguments{
\item{sce}{Milo object}

\item{genes}{Subset of rownames(sce) for which we will perform DE testing. Default = rownames(sce)}

\item{nhoods_sce}{Can be extracted from sce as nhoods(sce) prior to running the function. The reason the argument is passed by itself is to avoid calculating it every time.}

\item{hood.id}{Character specifying for which hood we should perform testing. Should be in colnames(nhoods_sce)}

\item{sample.id}{Character specifying which variable should be used as a sample/replica id. Should be in colData(sce)}

\item{condition.id}{Character specifying which variable should be used as a condition id. Should be in colData(sce)}

\item{covariates}{Vector specifying if additional covariates should be passed into experimental design. Default = NULL (no covariates)}

\item{min_n_cells_per_sample}{positive integer specifying the minimun number of cells per replica to be included in testing. Default = 2}

\item{gene_selection}{In {"all" , "per_hood"}. If = "per_hood", we will further discard genes not relevant for this particular hood.}

\item{min.count}{Positive integer, specifying min.count for gene selection. Default = 5}
}
\value{

}
\description{
Tests single hood for DE. Not intended to be used by itself (however possible), but rather as a part of \code{de_test_all_hoods}
}
