% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test_neighbourhoods.R
\name{de_test_single_neighbourhood}
\alias{de_test_single_neighbourhood}
\title{de_test_single_neighbourhood}
\usage{
de_test_single_neighbourhood(
  x,
  nhoods_sce,
  hood_id,
  sample_id,
  design,
  covariates,
  contrasts = NULL,
  min_n_cells_per_sample = 1,
  min_count = 3,
  run_separately = TRUE
)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object}

\item{nhoods_sce}{Can be extracted from x as \code{nhoods(x)} prior to running the function.}

\item{hood_id}{Character specifying for which hood we should perform testing. Should be in \code{colnames(nhoods_sce)}.}

\item{sample_id}{Character specifying which variable should be used as a replicate ID.
Should be in \code{colnames(colData(x))}. Default \code{sample_id = "sample"}.}

\item{design}{A \code{formula} object describing the experimental design for DE testing.
If \code{contrasts = NULL} (default), the last column column of model will be used for testing.}

\item{covariates}{Vector specifying all covariates that are passed into experimental design.
It should contain all columns used in the design formula (except \code{sample_id}).}

\item{contrasts}{NULL (default) or character string specifying what comparison to perform.
If you are unsure regarding the appropriate syntax for the \code{contrasts} in your data,
check \url{https://www.bioconductor.org/packages/release/bioc/vignettes/edgeR/inst/doc/edgeRUsersGuide.pdf}.

\emph{Note that at the moment we only support one comparison, if you wish to perform several comparisons, please run \code{de_test_neighbourhoods} for each comparison separately.}}

\item{min_n_cells_per_sample}{Positive integer specifying the minimum number of cells per replicate to be included in testing.
Default \code{min_n_cells_per_sample = 3}.}

\item{min_count}{Positive integer, specifying min.count for gene selection.
Default \code{min_count = 3}.}

\item{run_separately}{A boolean parameter specifying whether the function is to be run as a part of \code{\link{de_test_neighbourhoods}} (FALSE) or as a stand-alone run (TRUE). Default \code{run_separately = TRUE}.}
}
\value{
\code{data.frame} object containing miloDE results for the selected neighbourhood
}
\description{
Tests single hood for DE. Not intended to be used by itself (however possible), but rather as a part of \code{\link{`de_test_neighbourhoods`}}
}
\examples{
require(SingleCellExperiment)
require(miloR)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts =
floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") =
matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
nhoods_sce = nhoods(sce)
de_stat = de_test_single_neighbourhood(sce , nhoods_sce = nhoods_sce,
hood_id = 1 , sample_id = "sample" ,
design = ~type, covariates = c("type"))

}
