% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_not_perturbed_hoods.R
\name{filter_not_perturbed_hoods}
\alias{filter_not_perturbed_hoods}
\title{filter_not_perturbed_hoods}
\usage{
filter_not_perturbed_hoods(sce, min_cells = 3, auc.thresh = 0.5)
}
\arguments{
\item{sce}{sce_milo A Milo object (SCE + assigned hoods)}

\item{min_cells}{A positive integer specifying min number of cells in the hood for which we calculate AUC. Default = 3 (avoid changing this parameter).}

\item{auc.thresh}{A positive numeric specifying an AUC threshold below each we assign hoods to be no DE. Default = 0.5 (avoid changing this parameter)}
}
\value{

}
\description{
Pre-filtering hoods where we expect to have no DE (using classifier based method Augur that returns AUCs. AUC = 0.5)
}
