% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_pval_adjustment.R
\name{get_weights}
\alias{get_weights}
\title{get_weights}
\usage{
get_weights(nhoods_sce)
}
\arguments{
\item{nhoods_sce}{nhoods(sce)}
}
\value{

}
\description{
Calculate weights for spatial (across neighbourhoods) p-value adjustment.
}
\examples{
require(SingleCellExperiment)
require(miloR)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim" , k = 10 , order = 1)
nhoods_sce = nhoods(sce)
out = get_weights(nhoods_sce)
}
