% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_beeswarm_single_gene}
\alias{plot_beeswarm_single_gene}
\title{plot_beeswarm_single_gene}
\usage{
plot_beeswarm_single_gene(
  de_stat,
  gene,
  nhoodGroup,
  levels = NULL,
  alpha = 0.1,
  subset_nhoods = NULL,
  size = 2
)
}
\arguments{
\item{de_stat}{miloDE stat (output of \code{\link{de_test_neighbourhoods}}).
It does not have to be direct output of \code{\link{de_test_neighbourhoods}} i.e. it is allowed if \code{de_stat} has some additional assays/columns (e.g. calculated post hoc metrics on neighbourhoods/genes).}

\item{gene}{A character specifying the gene.}

\item{nhoodGroup}{A character specifying which values to use for neighbourhood grouping. Should be an assay in \code{de_stat} (or in \code{colnames(de_stat)} if \code{class(de_stat) == "data.frame"}).}

\item{levels}{NULL (default) or character vector specifying order for \code{nhoodGroup} values, in which they are to be plotted.
If \code{levels = NULL}, default order will be supplied.}

\item{alpha}{A numeric between 0 and 1 specifying the significance threshold. All neighbourhoods that are defined as not significant, will coloured in gray. Default \code{alpha = 0.1}.}

\item{subset_nhoods}{A vector (or NULL) specifying which neighbourhoods will be plotted.
Default \code{subset_nhoods = NULL} meaning that no subsetting is performed.}

\item{size}{A positive number specifying size of the dots. Default \code{size = 2}.}
}
\value{
Beeswarm plot, broke down by cell types; each point is a neighbourhood, colours and x-axis correspond to logFC (points are coloured if significant, based on \code{pval_corrected_across_nhoods}; otherwise in grey)
}
\description{
For the selected gene, returns a beeswarm plot, in which the DE statistics for the gene is binned by provided cell grouping (e.g. cell types);
}
\examples{

require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays =
list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") =
matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
de_stat = de_test_neighbourhoods(sce ,
design = ~type , covariates = c("type") )
de_stat$celltype = 1
p = plot_beeswarm_single_gene(de_stat ,
gene = "1" , nhoodGroup = "celltype")

}
