% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_mouseEmbryo.R
\docType{data}
\name{sce_mouseEmbryo}
\alias{sce_mouseEmbryo}
\title{MouseEmbryo, E8.5
Counts matrix for mouse embryo, E8.5 (data extracted from Pijuan-Sala et al., 2019): WT and chimera}
\format{
SCE object
}
\source{
MouseGastrulationData package
}
\usage{
data(sce_mouseEmbryo)
}
\description{
SCE object with log-normalised counts matrix strored in assay logcounts.
colData contains information about sample (i.e. batch), type (WT or chimera), tal1 (plus or minus).
}
