% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_pval_adjustment.R
\name{spatial_pval_adjustment}
\alias{spatial_pval_adjustment}
\title{spatial_pval_adjustment}
\usage{
spatial_pval_adjustment(weights = NULL, nhoods_sce, pvalues)
}
\arguments{
\item{nhoods_sce}{nhoods(sce)}

\item{pvalues}{vector of pvalues}
}
\value{

}
\description{
Performs p-values multiple testing correction, with accounting for the overlap. This is achieved by using weighted version of BH correction.
}
