% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_milo.R
\name{subset_milo}
\alias{subset_milo}
\title{subset_milo}
\usage{
subset_milo(sce_milo, colnames_2_retain)
}
\arguments{
\item{sce_milo}{Milo object}

\item{colnames_2_retain}{Character string of colnames to retain}
}
\value{

}
\description{
Function properly subsets data from sce-milo object (including nhoods(sce-milo))
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim.name = "reduced_dim" , k = 10 , order = 1)
sce = subset_milo(sce , colnames_2_retain = c("1", "2", "3"))
}
