% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_embedding.R
\name{add_embedding}
\alias{add_embedding}
\title{add_embedding}
\usage{
add_embedding(
  sce,
  genes = NULL,
  n_hvgs = 3000,
  assay_type = "logcounts",
  reduction_type = c("Azimuth", "MNN"),
  reducedDim_name,
  sample_id = "sample",
  cells_ref,
  cells_query,
  cell_id = NULL,
  d = 30
)
}
\arguments{
\item{sce}{SCE object}

\item{genes}{genes that should be used to calculate joint embedding (usually some sort of HVG selection). If = NULL (default) , we will calculate HVGs using scran::getTopHVGs}

\item{n_hvgs}{number of HVGs to retain for the joint embedding estimation}

\item{assay_type}{A string specifying the assay in SCE containing the expression values to be used (default = "logcounts")}

\item{reduction_type}{Either "Azimuth" or "MNN" - specifying which method to use for the embedding estimation}

\item{reducedDim_name}{String specifying the name under which to save the joint embedding}

\item{sample_id}{Character specifying which variable should be used as a sample/replica id. Should be in colData(sce)}

\item{cells_ref}{IDs of cells from reference samples}

\item{cells_query}{IDs of cells from query samples}

\item{cell_id}{Character specifying which variable should be used as a cell id (if colnames(sce) are not specified).}

\item{d}{Positive integer specifying number of dimensions in the joint embedding}
}
\value{

}
\description{
add_embedding
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
sce = SingleCellExperiment(assays = list( counts = matrix(rnorm(n_row*n_col), ncol=n_col)))
logcounts(sce) = matrix(rnorm(n_row*n_col), ncol=n_col)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
out = add_embedding(sce, reduction_type = "MNN" , reducedDim_name = "MNN", cells_ref = colnames(sce[, sce$type == "ref"]) ,  cells_query = colnames(sce[, sce$type == "query"]) , d = 5)
}
