% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_neighbourhoods.R
\name{assign_neighbourhoods}
\alias{assign_neighbourhoods}
\title{assign_neighbourhoods}
\usage{
assign_neighbourhoods(
  x,
  reducedDim_name,
  k = 25,
  prop = 0.2,
  order = 2,
  filtering = TRUE,
  k_init = 50,
  d = 30,
  verbose = TRUE
)
}
\arguments{
\item{x}{SingleCellExperiment object}

\item{reducedDim_name}{defines the slot in reducedDim(x) to use as embedding for graph construction}

\item{k}{Positive integer, defines how many neighbours to use for the hood assignment}

\item{prop}{Numerical, between 0 and 1, defines fraction of cells from SCE to use for the hoods}

\item{order}{In {1,2}, defines which order of neighbours to use}

\item{filtering}{In {TRUE,FALSE}, defines whether to filter hoods (reduces computing time greatly). Default = TRUE}

\item{k_init}{Positive integer, defines how many neighbours to use for identifying anchor cells}

\item{d}{Positive integer, defines how many dimensions from reducedDim(x) to use}

\item{verbose}{Boolean specifying whether to print intermediate output messages. Default = TRUE.}
}
\value{
Milo object containing cell-neighbourhood matrix in nhoods(out) slot
}
\description{
Assign hoods to SCE object
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays =
list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
reducedDim(sce , "reduced_dim") =
matrix(rnorm(n_col*n_latent), ncol=n_latent)
out = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
}
