% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_de_stat.R
\name{convert_de_stat}
\alias{convert_de_stat}
\title{convert_de_stat}
\usage{
convert_de_stat(de_stat)
}
\arguments{
\item{de_stat}{Output of milo-DE (from \code{de_test_all_neighbourhoods}), either in \code{data.frame} or \code{SingleCellExperiment}}
}
\value{

}
\description{
Converts output of milo-DE between \code{\link[base]{data.frame}} and \code{\linkS4class{SingleCellExperiment}} formats
}
\examples{
de_stat = expand.grid(gene = paste0("gene" , c(1:5)) , Nhood = c(1:10))
de_stat$Nhood_id = paste0("nhood_" , de_stat$Nhood)
de_stat$logFC = sample(seq(-2,2,1) , nrow(de_stat) , 1)
de_stat$pval = sample(c(0,1),nrow(de_stat),1)
de_stat$pval_corrected_across_genes = sample(c(0,1),nrow(de_stat),1)
de_stat$pval_corrected_across_nhoods = sample(c(0,1),nrow(de_stat),1)
de_stat$sufficient_n_samples = TRUE
de_stat$design_matrix_suitable = TRUE
de_stat = convert_de_stat(de_stat)
de_stat = convert_de_stat(de_stat)

}
