% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test_all_hoods.R
\name{de_test_all_hoods}
\alias{de_test_all_hoods}
\title{de_test_all_hoods}
\usage{
de_test_all_hoods(
  sce,
  sample_id = "sample",
  condition_id,
  discard_not_perturbed_hoods = c(TRUE, FALSE),
  covariates = NULL,
  min_n_cells_per_sample = 2,
  gene_selection = c("all", "none", "per_hood"),
  genes = rownames(sce),
  genes_2_exclude = NULL,
  min_count = 3
)
}
\arguments{
\item{sce}{Milo object}

\item{sample_id}{Character specifying which variable should be used as a sample/replica id. Should be in colData(sce)}

\item{condition_id}{Character specifying which variable should be used as a condition id. Should be in colData(sce)}

\item{discard_not_perturbed_hoods}{Boolean specifying whether perform prior hood selection using Augur based classifier. Note for big datasets it might take a while so recommended if a large portion of hoods is expected to be unperturbed.}

\item{covariates}{Vector specifying if additional covariates should be passed into experimental design. Default = NULL (no covariates)}

\item{min_n_cells_per_sample}{positive integer specifying the minimun number of cells per replica to be included in testing. Default = 2}

\item{gene_selection}{In {"none", "all" , "per_hood"}. If == "per_hood", it will further discard genes not relevant for this particular hood.
If == "all", it will perform overall selection and then test each hood across same genes (default). If "none" = no selection based on expression.}

\item{genes}{Character vector of genes for which we will perform DE testing. Default = rownames(sce). Note that it is upstream of gene selection (if applied).}

\item{genes_2_exclude}{Character vector of genes which we exlude from the selection. Default = NULL (none). Note that it is upstream of gene selection (if applied).}

\item{min_count}{Positive integer, specifying min.count for gene selection. Default = 3.}
}
\value{

}
\description{
Tests all hoods for DE + performs spatialFDR correction after
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_hoods(sce, reducedDim_name = "reduced_dim")
de_stat = de_test_all_hoods(sce , condition_id = "type" , gene_selection = "none", discard_not_perturbed_hoods = FALSE)
}
