% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test_neighbourhoods.R
\name{de_test_neighbourhoods}
\alias{de_test_neighbourhoods}
\title{de_test_neighbourhoods}
\usage{
de_test_neighbourhoods(
  x,
  sample_id = "sample",
  design,
  covariates,
  contrasts = NULL,
  subset_nhoods = NULL,
  min_n_cells_per_sample = 3,
  min_count = 3,
  output_type = "data.frame",
  plot_summary_stat = FALSE,
  layout = "UMAP",
  BPPARAM = NULL,
  correct_pvalues = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object.}

\item{sample_id}{Character specifying which variable should be used as a replicate ID.
Should be in \code{colnames(colData(x))}. Default \code{sample_id = "sample"}.}

\item{design}{A \code{formula} object describing the experimental design for DE testing.
Note that if \code{contrasts = NULL} (default), the last column column of model matrix will be used for testing.}

\item{covariates}{Vector specifying all covariates that are passed into experimental design.

\emph{It should contain all columns used in the design formula (except \code{sample_id})}.}

\item{contrasts}{NULL (default) or character string specifying what comparison to perform.
If you are unsure regarding the appropriate syntax for the \code{contrasts} in your data,
check \url{https://www.bioconductor.org/packages/release/bioc/vignettes/edgeR/inst/doc/edgeRUsersGuide.pdf}.

\emph{Note that at the moment we only support one comparison (i.e. one contrast); if you wish to perform several comparisons, please run \code{\link{de_test_neighbourhoods}} for each comparison separately.}}

\item{subset_nhoods}{NULL or character vector specifying the set of neighbourhoods that will be tested for DE.
Default \code{subset_nhoods = NULL} meaning no subsetting.}

\item{min_n_cells_per_sample}{Positive integer specifying the minimum number of cells per replicate to be included in testing.
Default \code{min_n_cells_per_sample = 3}.}

\item{min_count}{Positive integer, specifying \code{min.count} for gene selection (employes \code{\link[edgeR]{filterByExpr}}).
Default \code{min_count = 3}.}

\item{output_type}{In \code{c("data.frame","SCE")} Specifying the output format - either in \code{data.frame} or \code{\linkS4class{SingleCellExperiment}}.
Default \code{output_type = "data.frame"}.}

\item{plot_summary_stat}{Boolean specifying if we plot Milo neighbourhood plot summarising (per neighbourhood) whether testing was performed.
Default \code{plot_summary_stat = FALSE}.}

\item{layout}{A character indicating the name of the \code{reducedDim} slot in the \code{\linkS4class{Milo}} object to use for the layout of the plot(default \code{layout = "UMAP"}).
Only relevant if \code{plot_summary_stat = TRUE}.}

\item{BPPARAM}{NULL or \code{\link{MulticoreParam}} object to use for parallelisation (see \code{README} for the usage). Default \code{BPPARAM = NULL} meaning no parallelisation.
Note that if possible we recommend to parallel this in order to reduce computational time.}

\item{correct_pvalues}{Logical, whether to apply multiple testing correction across neighbourhoods
Default is TRUE. When FALSE, the 'pval_corrected_across_nhoods' assay will contain NA values to maintain compatibility with downstream plotting functions.}

\item{verbose}{Boolean specifying whether to print intermediate output messages. Default \code{verbose = TRUE}.}
}
\value{
\code{data.frame} or \code{SingleCellExperiment} object containing miloDE results for all supplied neighbourhoods.
For each tested gene-neighbourhood pair, we return logFC and p-values (raw and corrected across genes or neighbourhoods).
}
\description{
Performs DE testing within each neighbourhood + post hoc p-value correction across neighbourhoods. If a test for a gene x neighbourhood pair is not performed
(i.e. gene is not expressed in this neighbourhood and filtered out prior to testing), we returns NaNs.
}
\details{
We employ edgeR testing (using \code{\link[edgeR]{glmQLFit}}) within each neighbourhood.
We allow the user to submit the desired experimental design and incorporate various covariates, which is beneficial in the context of large cohort studies.
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts =
floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") =
matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
de_stat = de_test_neighbourhoods(sce , design = ~type,
covariates = c("type"), plot_summary_stat = FALSE)
de_stat = convert_de_stat(de_stat)
de_stat = convert_de_stat(de_stat)
}
