% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_test_neighbourhoods.R
\name{de_test_neighbourhoods}
\alias{de_test_neighbourhoods}
\title{de_test_neighbourhoods}
\usage{
de_test_neighbourhoods(
  x,
  sample_id = "sample",
  design,
  covariates,
  contrasts = NULL,
  subset_nhoods = NULL,
  min_n_cells_per_sample = 2,
  min_count = 3,
  output_type = "data.frame",
  plot_summary_stat = FALSE,
  layout = "UMAP",
  BPPARAM = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Milo object}

\item{sample_id}{Character specifying which variable should be used as a sample/replica id. Should be in colData(x)}

\item{design}{A \code{formula} object describing the experimental design for DE testing. If contrasts == NULL (default), the last column column of model will be used for testing.}

\item{covariates}{Vector specifying all covariates that are passed into experimental design. This should contain all columns used in the design formula.}

\item{contrasts}{NULL or contrast specifying what comparison to perform. Note that at the moment we only support one comparison, if you wish to perform several comparisons, please run \code{de_test_neighbourhoods} for each comparison separately.}

\item{subset_nhoods}{NULL or character vector specifying the set of neighbourhoods that will be tested for DE}

\item{min_n_cells_per_sample}{Positive integer specifying the minimum number of cells per replica to be included in testing. Default = 1.}

\item{min_count}{Positive integer, specifying min.count for gene selection. Default = 3.}

\item{output_type}{In {"data.frame","SCE"} Specifying the output format - either in data.frame or SingleCellExperiment (with assays corresponding to logFC, pvals (raw and corrected)); columns correspond to neighbourhoods. Note that default is data.frame, but if number of genes x neighbourhoods combinations is > 10^8 -- the output will be SCE.}

\item{plot_summary_stat}{Boolean specifying if we plot Milo neighbourhood plot summarising per neighbourhood whether testing was performed}

\item{layout}{A character indicating the name of the \code{reducedDim} slot in the \code{\linkS4class{Milo}} object to use for layout (default: 'UMAP'). Only relevant if plot_summary_stat == TRUE.}

\item{BPPARAM}{NULL or BPPARAM object to use for parallelization}

\item{verbose}{Boolean specifying whether to print intermediate output messages. Default = TRUE.}
}
\value{
data.frame (or SCE) object containing milo-DE results for all supplied neighbourhoods
}
\description{
Tests neighbourhoods for DE + performs per gene correction across neighbourhoods
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts =
floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") =
matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
de_stat = de_test_neighbourhoods(sce , design = ~type,
covariates = c("type"), plot_summary_stat = FALSE)
de_stat = convert_de_stat(de_stat)
de_stat = convert_de_stat(de_stat)
}
