% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_neighbourhood_sizes.R
\name{estimate_neighbourhood_sizes}
\alias{estimate_neighbourhood_sizes}
\title{estimate_neighbourhood_sizes}
\usage{
estimate_neighbourhood_sizes(
  sce,
  k_grid = seq(10, 100, 10),
  order = 2,
  prop = 0.1,
  filtering = TRUE,
  reducedDim_name,
  k_init = 50,
  d = 30,
  plot_stat = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sce}{SCE object}

\item{k_grid}{Vector of positive integers, defines how many neighbours to use for the hood assignment}

\item{order}{In {1,2}, defines which order of neighbours to use}

\item{prop}{Numerical, between 0 and 1, defines fraction of cells from SCE to use for the hoods. Default = 0.1.}

\item{filtering}{In {TRUE,FALSE}, defines whether to filter hoods (reduces computing time greatly). Default = TRUE}

\item{reducedDim_name}{defines the slot in reducedDim(sce) to use as embedding for graph construction}

\item{k_init}{Positive integer, defines how many neighbours to use for identifying anchor cells}

\item{d}{Positive integer, defines how many dimensions from reducedDim(sce) to use}

\item{plot_stat}{Boolean specifying whether to plot the stat}

\item{verbose}{Boolean specifying whether to print intermediate output messages. Default = FALSE.}
}
\value{

}
\description{
For a grid of k (order fixed), return hood size distribution; that will help a user to select appropriate k
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
out = estimate_neighbourhood_sizes(sce, k_grid = c(5,10), reducedDim_name = "reduced_dim")

}
