% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_neighbourhoods.R
\name{filter_neighbourhoods}
\alias{filter_neighbourhoods}
\title{filter_neighbourhoods}
\usage{
filter_neighbourhoods(sce_milo)
}
\arguments{
\item{sce_milo}{A Milo object (SCE + assigned hoods)}
}
\value{

}
\description{
Filtering redundant hoods, using the greedy approach to set cover problem
}
\examples{

}
