% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_neighbourhoods.R
\name{filter_neighbourhoods}
\alias{filter_neighbourhoods}
\title{filter_neighbourhoods}
\usage{
filter_neighbourhoods(sce_milo)
}
\arguments{
\item{sce_milo}{A Milo object (SCE + assigned hoods)}
}
\value{

}
\description{
Filtering redundant hoods, using the greedy approach to set cover problem
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim.name = "reduced_dim" , k = 10 , order = 1)
sce = filter_neighbourhoods(sce)
}
