% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_beeswarm_gene_set}
\alias{plot_beeswarm_gene_set}
\title{plot_beeswarm_gene_set}
\usage{
plot_beeswarm_gene_set(
  de_stat,
  genes,
  nhoodGroup,
  logFC_correction = TRUE,
  correction_by = "pval_corrected_across_nhoods",
  alpha = 0.1,
  subset_nhoods = NULL,
  size = 2
)
}
\arguments{
\item{de_stat}{milo-DE stat (output of \code{\link[miloDE]{de_test_neighbourhoods}})}

\item{genes}{A character specifying genes ID}

\item{nhoodGroup}{A character specifying which column to use for neighbourhood grouping.}

\item{logFC_correction}{Boolean specifying whether to perfrom logFC correction. If TRUE (default), logFC will be set to 0 if corrected pvalue (defined by 'correction_by' variable) < alpha}

\item{correction_by}{Character - colname from de_stat - specifying which column to use to decide on significance for logFC. Relevant only if logFC_correction = TRUE.}

\item{alpha}{A numeric between 0 and1 specifying the significance threshold. Default = 0.1.}

\item{subset_nhoods}{NULL or numeric vector specifying which Nhoods to use}

\item{size}{A positive number specifying size of the dots}
}
\value{
beeswarm, broke down by cell types; each point is a neighbourhood, colour - average logFC across selected genes; x - fraction of genes that are DE
}
\description{
Returns beeswarm plot for many genes
}
\examples{

require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts =
floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") =
matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
de_stat = de_test_neighbourhoods(sce , design = ~type ,
covariates = c("type") )
de_stat$celltype = 1
p = plot_beeswarm_gene_set(de_stat , genes = c("1","2") ,
nhoodGroup = "celltype")

}
