% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_milo_by_single_metric}
\alias{plot_milo_by_single_metric}
\title{plot_milo_by_single_metric}
\usage{
plot_milo_by_single_metric(
  sce,
  nhood_stat,
  colour_by = "logFC",
  significance_by = NULL,
  order_by = NULL,
  order_direction = TRUE,
  size_by = NULL,
  alpha = 0.1,
  layout = "UMAP",
  subset_nhoods = NULL,
  size_range = c(1, 3),
  node_stroke = 0.3,
  edge_width = c(0.2, 0.5),
  edge_weight.thresh = NULL
)
}
\arguments{
\item{sce}{\code{\linkS4class{Milo}} object}

\item{nhood_stat}{data.frame object, containing columns 'Nhood' (should correspond to neighbourhoods from nhoodGraph(sce))}

\item{colour_by}{A colname from nhood_stat - nodes will be coloured by the values from this column}

\item{significance_by}{A colname from nhood_stat (or NULL) - if values for this column exceed alpha, colour_by will be set to 0. Default = NULL meaning that we will not use any column to define the significance (i.e. no correction for colour_by)}

\item{order_by}{A colname from nhood_stat (or NULL) specifying by which column we will order neighbourhoods for plotting. Default = NULL meaning we will use size of the neighbourhoods.}

\item{order_direction}{Boolean specifying the direction of ordering neighbourhoods}

\item{size_by}{NULL or a character vector - colname from nhood_stat - specifying metric by which we define size for neighbourhoods}

\item{alpha}{A scalar (between 0 and 1) specifying the significance level used}

\item{layout}{A character indicating the name of the \code{reducedDim} slot in the \code{\linkS4class{Milo}} object to use for layout (default: 'UMAP')}

\item{subset_nhoods}{A vector (or NULL) specifying which neighbourhoods will be plotted. Default = NULL meaning that all neighbourhoods will be plotted}

\item{size_range}{A numeric vector indicating the range (min and max) of node sizes to use for plotting (to avoid overplotting in the graph)}

\item{node_stroke}{A numeric indicating the desired thickness of the border around each node}

\item{edge_width}{A numeric vector indicating the range (min and max) of edge widths to use for plotting}

\item{edge_weight.thresh}{A numeric (or NULL) specifying a threshold for minimum cells in common (between neighbourhoods) required for an edge to be plotted}
}
\value{

}
\description{
Return Milo-graph plot; each node is coloured by colour_by column from nhood_stat, if significance_by is smaller than alpha. Vetices are ordered by order_by column
}
\examples{
require(SingleCellExperiment)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$sample = floor(runif(n = n_col , min = 1 , max = 5))
sce$type = ifelse(sce$sample \%in\% c(1,2) , "ref" , "query")
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim_name = "reduced_dim")
de_stat = de_test_neighbourhoods(sce , condition_id = "type" , gene_selection = "none")
de_stat = de_stat[de_stat$gene == "1", ]
umaps = as.data.frame(matrix(rnorm(n_col*2), ncol=2))
colnames(umaps) = c("V1" , "V2")
reducedDim(sce , "UMAP") = umaps
p = plot_milo_by_single_metric(sce, de_stat)

}
