% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_neighbourhoods_by_DE_magnitude.R
\name{rank_neighbourhoods_by_DE_magnitude}
\alias{rank_neighbourhoods_by_DE_magnitude}
\title{rank_neighbourhoods_by_DE_magnitude}
\usage{
rank_neighbourhoods_by_DE_magnitude(de_stat, pval.thresh = 0.1, z.thresh = -3)
}
\arguments{
\item{de_stat}{Output of milo-DE, either in data.frame format or SCE}

\item{pval.thresh}{A scalar specifying which threshold to use for deciding on significance for gene being DE in a hood. Default = 0.1.}

\item{z.thresh}{A scalar specifying which threshold to use for deciding on which z-normalised p-values are going to be considered specifically DE. Default = -3.}
}
\value{

}
\description{
Ranks neighbourhoods by the magnitude of DE: number of DE genes and number of 'specifically' DE genes. Note that for this analysis we set NaN p-values (raw and corrected) to 1 - interpret accordingly.
}
\examples{
de_stat = expand.grid(gene = paste0("gene" , c(1:5)) , Nhood = c(1:10))
de_stat$Nhood_center = paste0("nhood_" , de_stat$Nhood)
de_stat$logFC = sample(seq(-2,2,1) , nrow(de_stat) , 1)
de_stat$pval = sample(c(0,1),nrow(de_stat),1)
de_stat$pval_corrected_across_genes = sample(c(0,1),nrow(de_stat),1)
de_stat$pval_corrected_across_nhoods = sample(c(0,1),nrow(de_stat),1)
de_stat$sufficient_n_samples = TRUE
de_stat$design_matrix_suitable = TRUE
out = rank_neighbourhoods_by_DE_magnitude(de_stat)

}
