% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_mouseEmbryo.R
\docType{data}
\name{sce_mouseEmbryo}
\alias{sce_mouseEmbryo}
\title{Chimeric mouse embryo, Tal1- (Pijuan-Sala et al., 2019)}
\format{
\code{\linkS4class{SingleCellExperiment}} object
}
\source{
MouseGastrulationData package
}
\usage{
data(sce_mouseEmbryo)
}
\description{
\code{SingleCellExperiment} object, containing counts
(raw are pulled using \code{\link[MouseGastrulationData]{Tal1ChimeraData}}
and log-normalised are estimated using \code{\link[scuttle]{logNormCounts}}) matrices
for chimeric mouse embryos (both Tal1+ and Tal1-).
}
\details{
Additionally, we subselected only 3 cell types (Endothelium , Blood progenitors 2, Neural crest)
and 300 HVGs.

\code{colData(sce_mouseEmbryo)} contains information about replicate (\code{sample}) and Tal1 status (\code{tomato = TRUE} means Tal1-).
}
