% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_pval_adjustment.R
\name{spatial_pval_adjustment}
\alias{spatial_pval_adjustment}
\title{spatial_pval_adjustment}
\usage{
spatial_pval_adjustment(nhoods_x, pvalues)
}
\arguments{
\item{nhoods_x}{Should be extracted from x as \code{nhoods(x)}.}

\item{pvalues}{Vector of p-values.}
}
\value{
Vector with \sQuote{spatially} (i.e. across neighbourhoods) adjusted p-values
}
\description{
Performs p-values multiple testing correction, with accounting for the overlap
}
\details{
This function is not intended to be run by itself, but can be useful if the user wants to perform \sQuote{spatially aware} multiple testing correction.

Under the hood it performs weighted version of BH correction, where weights are reciprocal to the local desnity of a neighbourhood.
Accordingly, big and/or highly connected neighbourhoods will have lower weight.
}
\examples{
require(SingleCellExperiment)
require(miloR)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts =
floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
reducedDim(sce , "reduced_dim") =
matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce,
reducedDim_name = "reduced_dim" , k = 10 , order = 1)
nhoods_x = nhoods(sce)
pvalues = runif(n = ncol(nhoods_x) , min = 0 , max = 1)
out = spatial_pval_adjustment(nhoods_x, pvalues = pvalues)
}
