% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_pval_adjustment.R
\name{spatial_pval_adjustment}
\alias{spatial_pval_adjustment}
\title{spatial_pval_adjustment}
\usage{
spatial_pval_adjustment(weights = NULL, nhoods_sce, pvalues)
}
\arguments{
\item{nhoods_sce}{nhoods(sce)}

\item{pvalues}{vector of pvalues}
}
\value{

}
\description{
Performs p-values multiple testing correction, with accounting for the overlap. This is achieved by using weighted version of BH correction.
}
\examples{
require(SingleCellExperiment)
require(miloR)
n_row = 500
n_col = 100
n_latent = 5
sce = SingleCellExperiment(assays = list(counts = floor(matrix(rnorm(n_row*n_col), ncol=n_col)) + 4))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
reducedDim(sce , "reduced_dim") = matrix(rnorm(n_col*n_latent), ncol=n_latent)
sce = assign_neighbourhoods(sce, reducedDim.name = "reduced_dim" , k = 10 , order = 1)
nhoods_sce = nhoods(sce)
pvalues = runif(n = ncol(nhoods_sce) , min = 0 , max = 1)
out = spatial_pval_adjustment(weights = NULL , nhoods_sce = nhoods_sce, pvalues = pvalues)
}
